Rename protocol.py to old_protocol.py
