+++ src/berconpy/__init__.py
@@ -3,7 +3,7 @@ from .client import AsyncRCONClient
 from .errors import *
 from .packet import *
 from .player import Player
-from .protocol import RCONClientDatagramProtocol
+from .old_protocol import RCONClientDatagramProtocol
 from . import utils
 
 __version__ = "1.1.0"
